made the ambassador card component
