+++ app/globals.css
@@ -1675,7 +1675,7 @@ input {
     #9f32f8,
     #bf97df
   ); */
-  background-color: #7300d0;
+  background-color: rgb(144, 0, 255)  ;
   padding: 10px 20px;
   background-size: 200% 200%;
   border-radius: 0.5rem;
@@ -1689,9 +1689,9 @@ input {
 }
 
 .signInbut:hover {
-  box-shadow: 0px 0px 5px rgb(144, 0, 255);
+  box-shadow: 0px 0px 5px #7300d0;
   /* animation: ani 7s infinite linear; */
-  background-color: rgb(144, 0, 255);
+  background-color: #7300d0;
 }
 
 /* ! edawg */
